Move workload annotation mapping logic to selector
